// src/app/types/AudioName.ts
export enum AudioName {
  EASY_LOOP = 'EASY_LOOP',
  MEDIUM_LOOP = 'MEDIUM_LOOP',
  HARD_LOOP = 'HARD_LOOP',
  CORRECT_ANSWER = 'CORRECT_ANSWER',
  WRONG_ANSWER = 'WRONG_ANSWER',
  FINAL_ANSWER = 'FINAL_ANSWER',
  COPMMERCIAL_BREAK = 'COPMMERCIAL_BREAK',
  PHONE_CALL = 'PHONE_CALL',
  LETS_PLAY = 'LETS_PLAY',
  MAIN = 'MAIN',
  QUICK_MIND = 'QUICK_MIND',
}